Feature/streamliningandfixes

+++ CustomerPurchases/Controllers/PurchasesController.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Net.Http;
 using CustomerPurchases.Data;
 using CustomerPurchases.Data.Products;
 using CustomerPurchases.Models;
@@ -6,6 +8,9 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using System.Threading.Tasks;
+using CustomerPurchases.Models.DTOs;
+using Microsoft.Extensions.Configuration;
+using System.Linq;
 
 namespace CustomerPurchases.Controllers
 {
@@ -13,11 +18,15 @@ namespace CustomerPurchases.Controllers
     {
         private readonly IPurchaseRepo _repository;
         private readonly IProductService _productServ;
+        private readonly IHttpClientFactory _clientFactory;
+        private readonly IConfiguration _config;
 
-        public PurchasesController(IPurchaseRepo repository, IProductService prodServ)
+        public PurchasesController(IPurchaseRepo repository, IProductService prodServ, IHttpClientFactory factory, IConfiguration config)
         {
             _repository = repository;
             _productServ = prodServ;
+            _clientFactory = factory;
+            _config = config;
         }
 
         // GET: Purchases
@@ -28,7 +37,7 @@ namespace CustomerPurchases.Controllers
         }
 
         // GET: Purchases/Details/5
-        public async Task<IActionResult> Details(int? id)
+        public async Task<ActionResult<PurchaseDetailsDto>> Details(int? id)
         {
             if (id == null)
             {
@@ -41,17 +50,22 @@ namespace CustomerPurchases.Controllers
                 return NotFound();
             }
 
-            return View(purchase);
+            return Ok(new PurchaseDetailsDto
+            {
+                Id = purchase.Id,
+                AccountId = purchase.AccountId,
+                AddressId = purchase.AddressId,
+                OrderStatus = purchase.OrderStatus,
+                ProductId = purchase.ProductId,
+                Qty = purchase.Qty,
+                TimeStamp = purchase.TimeStamp
+            });
         }
 
         // GET: Purchases/Create
         public async Task<IActionResult> Create()
         {
-            // TODO - Check stock, Address, Phone no
             ViewData["ProductId"] = new SelectList(await _productServ.GetAll(), "Id", "Id");
-            // TODO - Populate these viewbags
-            ViewData["AddressId"] = null;
-            ViewData["AccountId"] = null;
 
             return View();
         }
@@ -63,13 +77,26 @@ namespace CustomerPurchases.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,ProductId,Qty,AddressId,AccountId,OrderStatus")] Purchase purchase)
         {
-            // TODO - Check stock, Address, Phone no
             if (ModelState.IsValid)
             {
-                purchase.TimeStamp = DateTime.Now;
-                _repository.InsertPurchase(purchase);
-                await _repository.Save();
-                return RedirectToAction(nameof(Index));
+                // TODO - Check Stock
+                var client = _clientFactory.CreateClient("RetryAndBreak");
+                client.BaseAddress = new System.Uri(_config["StockURL"]);
+                var resp = await client.GetAsync("api/stock/");
+
+                if (resp.IsSuccessStatusCode)
+                {
+                    var product = await resp.Content.ReadAsAsync<ProductStockDto>();
+                    if (product.Stock <= 0)
+                    {
+                        return BadRequest("Product is out of stock");
+                    }
+                    // TODO - Check Account info (Address, Phone)
+                    purchase.TimeStamp = DateTime.Now;
+                    _repository.InsertPurchase(purchase);
+                    await _repository.Save();
+                    return RedirectToAction(nameof(Index));
+                }
             }
             ViewData["ProductId"] = new SelectList(await _productServ.GetAll(), "Id", "Id", purchase.ProductId);
             return View(purchase);
@@ -150,10 +177,41 @@ namespace CustomerPurchases.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (id == null || id < 0)
+            {
+                return BadRequest();
+            }
             _repository.DeletePurchase(id);
             await _repository.Save();
             return RedirectToAction(nameof(Index));
         }
+
+        public async Task<ActionResult<List<Purchase>>> OrderHistory(int accId)
+        {
+            if (accId < 0)
+            {
+                return BadRequest();
+            }
+            var purchases = await _repository.GetPurchaseByAccount(accId);
+            if (purchases.Any())
+            {
+                return Ok(purchases);
+            }
+
+            return NotFound();
+        }
+        
+        // TODO - Implement separately
+        //public async Task<IActionResult> OrderHistory(int accId)
+        //{
+        //    var purchases = await _repository.GetPurchaseByAccount(accId);
+        //    if (purchases.Any())
+        //    {
+        //        return View("OrderHistory",purchases);
+        //    }
+
+        //    return View();
+        //}
 
         private async Task<bool> PurchaseExists(int id)
         {
